Added Mirages
Mirages are images of where a unit will be after an order is performed.
They are implemented as a struct attached to every order. Every frame,
mirages are drawn with the same sprite as the unit, but with a
blue-transparent tint.

Problems:
* Attack orders generate unnecessary mirages, causing weird things with
overlap. This may not be an issue after attack actions are given separate
sprites.
* More stuff about things being improperly hosted in PlayerController.
* Mirages aren't ordered, you can't tell which ones go in what order.
Could be fixed by rendering a number next to the sprite, or (my favorite)
adding a trail that goes from the base unit to mirage to mirage.